[wip] lapack

+++ src/include/sbdsqr.h
@@ -1,8 +1,9 @@
 pure subroutine SBDSQR( UPLO, N, NCVT, NRU, NCC, D, E, VT, LDVT, U, &
                &        LDU, C, LDC, WORK, INFO )
-CHARACTER,intent(in) :: UPLO
-INTEGER,intent(in)   :: LDC, LDU, LDVT, N, NCC, NCVT, NRU
-INTEGER,intent(out)  :: INFO
-real,intent(inout)   :: C( LDC, * ), D( * ), E( * ), U( LDU, * ), VT( LDVT, * )
-real,intent(out)     :: WORK( * )
+use, intrinsic :: ISO_FORTRAN_ENV, only: SP => REAL32
+CHARACTER,intent(in)   :: UPLO
+INTEGER,intent(in)     :: LDC, LDU, LDVT, N, NCC, NCVT, NRU
+INTEGER,intent(out)    :: INFO
+real(SP),intent(inout) :: C( LDC, * ), D( * ), E( * ), U( LDU, * ), VT( LDVT, * )
+real(SP),intent(out)   :: WORK( * )
 end subroutine SBDSQR
