+++ src/include/sorgbr.h
@@ -1,8 +1,9 @@
 pure subroutine SORGBR( VECT, M, N, K, A, LDA, TAU, WORK, LWORK, INFO )
-character,intent(in) :: VECT
-INTEGER,intent(in)   :: K, LDA, LWORK, M, N
-INTEGER,intent(out)  :: INFO
-real,intent(in)      :: TAU( * )
-real,intent(inout)   :: A( LDA, * )
-real,intent(out)     :: WORK( * )
+use, intrinsic :: ISO_FORTRAN_ENV, only: SP => REAL32
+character,intent(in)   :: VECT
+INTEGER,intent(in)     :: K, LDA, LWORK, M, N
+INTEGER,intent(out)    :: INFO
+real(SP),intent(in)    :: TAU( * )
+real(SP),intent(inout) :: A( LDA, * )
+real(SP),intent(out)   :: WORK( * )
 end subroutine SORGBR
